requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
eventlet==0.33.3
inquirer==3.4.1
google-generativeai>=0.3.0
